+++ mzmine-community/src/main/java/io/github/mzmine/modules/tools/output_compare_csv/OutputCheckResultsTab.java
@@ -1,0 +1,15 @@
+package io.github.mzmine.modules.tools.output_compare_csv;
+
+import io.github.mzmine.gui.mainwindow.SimpleTab;
+import java.util.List;
+
+public class OutputCheckResultsTab extends SimpleTab {
+
+
+  public OutputCheckResultsTab(final String title, List<DataCheckResult> checks) {
+    super(title);
+
+//    FxTextFlows.newTextFlow()
+//    FxTexts.styledText()
+  }
+}
